Feat(triggers): extended the candidate date for triggers
Data that is required for voltage buffers is now included. 

Include only the necessary data for each candidate list - make the code slightly easier to manage.

Fixed an error for boxcar clustering where DM limit was using time head time meaning the data was really just clustered in time and not DM.