Feat(cluster): added function for delta DM
Function that calculates an approximation for the DM offset that lowers the candidate SNR to the requested level. This value will be used to draw a box around the candidate for clustering. It will be calculated for every candidate considered for clustering (both TB and final clustering). No need to calculate it for non-clustered candidates. Value currently not used (relevant code commented out), as its addition breaks the dictionary keys that are obtained dynamically from the candidate dictionary.